refactor: KPI retrieval methods for courses and actions; update API endpoints in frontend services
