+++ NERBABO.Backend/NERBABO.ApiService/Core/Kpis/Services/KpisService.cs
@@ -479,4 +479,64 @@ public class KpisService(
 
         return (startDate, refersTo);
     }
+
+    public async Task<Result<KpisCourseDto>> GetCourseKpisAsync(long courseId)
+    {
+        var existingCourse = await _context.Courses
+            .Include(c => c.Actions)
+                .ThenInclude(a => a.ActionEnrollments)
+                    .ThenInclude(ae => ae.Participations)
+            .Include(c => c.Actions)
+                .ThenInclude(a => a.ModuleTeachings)
+                    .ThenInclude(mt => mt.Sessions)
+            .FirstOrDefaultAsync(c => c.Id == courseId);
+
+        if (existingCourse is null)
+        {
+            _logger.LogWarning("Course with ID {id} not found.", courseId);
+            return Result<KpisCourseDto>
+                .Fail("Não encontrado.", "Curso não encontrado.",
+                StatusCodes.Status404NotFound);
+        }
+
+        // Calculate aggregated KPIs from all actions of this course
+        var kpis = new KpisCourseDto
+        {
+            TotalStudents = existingCourse.Actions.Sum(a => a.TotalStudents),
+            TotalApproved = existingCourse.Actions.Sum(a => a.TotalApproved),
+            TotalVolumeHours = existingCourse.Actions.Sum(a => a.TotalVolumeHours),
+            TotalVolumeDays = existingCourse.Actions.Sum(a => a.TotalVolumeDays)
+        };
+
+        return Result<KpisCourseDto>
+            .Ok(kpis);
+    }
+
+    public async Task<Result<KpisActionDto>> GetActionKpisAsync(long actionId)
+    {
+        var existingAction = await _context.Actions
+            .Include(a => a.ActionEnrollments)
+                .ThenInclude(ae => ae.Participations)
+            .Include(a => a.ModuleTeachings).ThenInclude(mt => mt.Sessions)
+            .FirstOrDefaultAsync(a => a.Id == actionId);
+
+        if (existingAction is null)
+        {
+            _logger.LogWarning("CourseAction with ID {id} not found.", actionId);
+            return Result<KpisActionDto>
+                .Fail("Não encontrado.", "Ação não encontrada.",
+                StatusCodes.Status404NotFound);
+        }
+
+        var kpis = new KpisActionDto
+        {
+            TotalStudents = existingAction.TotalStudents,
+            TotalApproved = existingAction.TotalApproved,
+            TotalVolumeDays = existingAction.TotalVolumeDays,
+            TotalVolumeHours = existingAction.TotalVolumeHours
+        };
+
+        return Result<KpisActionDto>
+            .Ok(kpis);
+    }
 }
